(#550) Add sensu client de-registration
Without this patch the sensu client does not support the deregister and
deregistration attributes.  These attributes are documented at [deregistration
attributes](https://sensuapp.org/docs/latest/reference/clients#deregistration-attributes)

This patch adds the `sensu::client_deregister` and
`sensu::client_deregistration` class parameters which flow through to the
sensu_client_config type and provider.

This patch also refactors the sensu_client_config type and provider to operate
the same way as the sensu_check type and provider.  Specifically, the JSON
provider dynamically introspects the set of properties from the type and
creates getter and setter methods to output the JSON.  Custom attributes are
enhanced, if `client_custom = absent` then all custom attributes in the client
configuration are removed.  Otherwise, the custom attributes are merged on top
of any existing custom attributes.

When the client config is flushed to /etc/sensu/conf.d/client.json the top
level set of keys are sorted.  This isn't terribly useful as there is usually
just one key, `{"client":{}}`.  The value of the client attribute is also
sorted, making the output more useful and human readable.

Resolves #550